Added 510 - Not Extended HTTP Response Phrase
I added the 510 - Not Extended HTTP Response Phrase, as defined on the [RFC2774](https://tools.ietf.org/html/rfc2774#section-7).

Should it be merged? Maybe.

RFC2774 status is "Experimental" since February 2000, but I haven't found any conflict with other status codes. I'd say it'd be better to use it since some servers might respond with this status (I'm dealing with an API which does, and got quite confuse with an empty reason phrase on my logs).